dec_base/requirements.txt
-r ../restless/requirements.txt
django-uni-form==0.9.0
django-crispy-forms==1.1.2
django-taggit==0.9.3
django-extensions==0.8
django-templatetag-sugar==0.1
django-pagination==1.0.7
django-bootstrap-pagination==0.1.10
django-paging==0.2.4
django-indexer==0.3.0
django-debug-toolbar==0.9.4
django-selectable==0.5.2
http://ge.dec.wa.gov.au/_/pypi/django-tinymce-f3d4b92.tar.gz
docutils==0.8.1
tablib==0.9.11
Pycco==0.3.0
selenium==2.21.1
Unidecode==0.04.9
